refactor: restructuring code
refactor to allow easily extending functionality i.e. extending palindromechecker to analyze sentences and so on

danielherrerohernando/ohce-kata#4
