--- src/inputHandler.js
@@ -1,7 +1,0 @@
-const reverseInput = input =>
-	input
-		.split('')
-		.reverse()
-		.join('');
-
-module.exports = { reverseInput };
