+++ src/lib/stdoutWriter.js
@@ -1,0 +1,3 @@
+const write = data => data && process.stdout.write(data + '\n');
+
+module.exports = { write };
