+++ src/lib/timeHandler.js
