+++ test/unit/greetings.test.js
@@ -1,5 +1,5 @@
 const { expect } = require('chai');
-const { greetUser, goodbyeUser } = require('../../src/greetings');
+const { greetUser, goodbyeUser } = require('../../src/lib/greetings');
 
 describe('Unit test - Greetings Handler', () => {
 	it('Should greet user including the username', () => {
